Correção Criar Despesas.xml

+++ app/src/main/java/br/edu/infnet/meusgastos/models/Despesa.kt
@@ -4,7 +4,7 @@ import java.time.LocalDate
 
 data class Despesa(
     val nome: String = "",
-    val valor: Double = 0.0,
+    val valor: Float = 0F,
     val data: LocalDate,
     val descricao: String = "",
     val categoriaId: Int = 0
